First pass on layered mappings, not tested at all.
