+++ src/main/java/net/fabricmc/loom/LoomGradleExtension.java
@@ -52,10 +52,11 @@ import net.fabricmc.loom.configuration.LoomProjectData;
 import net.fabricmc.loom.configuration.ide.RunConfigSettings;
 import net.fabricmc.loom.configuration.processors.JarProcessor;
 import net.fabricmc.loom.configuration.processors.JarProcessorManager;
-import net.fabricmc.loom.configuration.providers.MinecraftProvider;
-import net.fabricmc.loom.configuration.providers.mappings.MappingsProvider;
-import net.fabricmc.loom.configuration.providers.mappings.mojmap.MojangMappingsDependency;
-import net.fabricmc.loom.configuration.providers.mappings.mojmap.MojangMappingsSpec;
+import net.fabricmc.loom.configuration.providers.MinecraftProviderImpl;
+import net.fabricmc.loom.configuration.providers.mappings.GradleMappingContext;
+import net.fabricmc.loom.configuration.providers.mappings.LayeredMappingSpecBuilder;
+import net.fabricmc.loom.configuration.providers.mappings.LayredMappingsDependency;
+import net.fabricmc.loom.configuration.providers.mappings.MappingsProviderImpl;
 import net.fabricmc.loom.configuration.providers.minecraft.MinecraftMappedProvider;
 
 public class LoomGradleExtension {
@@ -115,13 +116,13 @@ public class LoomGradleExtension {
 	}
 
 	public Dependency officialMojangMappings() {
-		return new MojangMappingsDependency(project, this, new MojangMappingsSpec());
+		return layered(LayeredMappingSpecBuilder::officalMojangMappings);
 	}
 
-	public Dependency officialMojangMappings(Action<MojangMappingsSpec> mojangMappingsSpecAction) {
-		var mojangMappingsSpec = new MojangMappingsSpec();
-		mojangMappingsSpecAction.execute(mojangMappingsSpec);
-		return new MojangMappingsDependency(project, this, mojangMappingsSpec);
+	public Dependency layered(Action<LayeredMappingSpecBuilder> action) {
+		LayeredMappingSpecBuilder builder = new LayeredMappingSpecBuilder();
+		action.execute(builder);
+		return new LayredMappingsDependency(new GradleMappingContext(project), builder.build());
 	}
 
 	public LoomGradleExtension(Project project) {
@@ -253,7 +254,7 @@ public class LoomGradleExtension {
 			return new File((String) project.property("fabric.loom.natives.dir"));
 		}
 
-		File natives = new File(getUserCache(), "natives/" + getMinecraftProvider().getMinecraftVersion());
+		File natives = new File(getUserCache(), "natives/" + getMinecraftProvider().minecraftVersion());
 
 		if (!natives.exists()) {
 			natives.mkdirs();
@@ -278,16 +279,16 @@ public class LoomGradleExtension {
 		return dependencyManager;
 	}
 
-	public MinecraftProvider getMinecraftProvider() {
-		return getDependencyManager().getProvider(MinecraftProvider.class);
+	public MinecraftProviderImpl getMinecraftProvider() {
+		return getDependencyManager().getProvider(MinecraftProviderImpl.class);
 	}
 
 	public MinecraftMappedProvider getMinecraftMappedProvider() {
 		return getMappingsProvider().mappedProvider;
 	}
 
-	public MappingsProvider getMappingsProvider() {
-		return getDependencyManager().getProvider(MappingsProvider.class);
+	public MappingsProviderImpl getMappingsProvider() {
+		return getDependencyManager().getProvider(MappingsProviderImpl.class);
 	}
 
 	public void setDependencyManager(LoomDependencyManager dependencyManager) {
@@ -349,7 +350,7 @@ public class LoomGradleExtension {
 	// Creates a new file each time its called, this is then held onto later when remapping the output jar
 	// Required as now when using parallel builds the old single file could be written by another sourceset compile task
 	public synchronized File getNextMixinMappings() {
-		File mixinMapping = new File(getProjectBuildCache(), "mixin-map-" + getMinecraftProvider().getMinecraftVersion() + "-" + getMappingsProvider().mappingsVersion + "." + mixinMappings.size() + ".tiny");
+		File mixinMapping = new File(getProjectBuildCache(), "mixin-map-" + getMinecraftProvider().minecraftVersion() + "-" + getMappingsProvider().mappingsVersion + "." + mixinMappings.size() + ".tiny");
 		mixinMappings.add(mixinMapping);
 		return mixinMapping;
 	}
