+++ src/main/java/net/fabricmc/loom/configuration/providers/minecraft/MinecraftMappedProvider.java
@@ -36,8 +36,8 @@ import com.google.common.collect.ImmutableMap;
 import org.gradle.api.Project;
 
 import net.fabricmc.loom.configuration.DependencyProvider;
-import net.fabricmc.loom.configuration.providers.MinecraftProvider;
-import net.fabricmc.loom.configuration.providers.mappings.MappingsProvider;
+import net.fabricmc.loom.configuration.providers.MinecraftProviderImpl;
+import net.fabricmc.loom.configuration.providers.mappings.MappingsProviderImpl;
 import net.fabricmc.loom.util.Constants;
 import net.fabricmc.loom.util.TinyRemapperMappingsHelper;
 import net.fabricmc.tinyremapper.OutputConsumerPath;
@@ -53,7 +53,7 @@ public class MinecraftMappedProvider extends DependencyProvider {
 	private File minecraftMappedJar;
 	private File minecraftIntermediaryJar;
 
-	private MinecraftProvider minecraftProvider;
+	private MinecraftProviderImpl minecraftProvider;
 
 	public MinecraftMappedProvider(Project project) {
 		super(project);
@@ -101,7 +101,7 @@ public class MinecraftMappedProvider extends DependencyProvider {
 	private void mapMinecraftJar() throws IOException {
 		String fromM = "official";
 
-		MappingsProvider mappingsProvider = getExtension().getMappingsProvider();
+		MappingsProviderImpl mappingsProvider = getExtension().getMappingsProvider();
 
 		Path input = minecraftProvider.getMergedJar().toPath();
 		Path outputMapped = minecraftMappedJar.toPath();
@@ -148,7 +148,7 @@ public class MinecraftMappedProvider extends DependencyProvider {
 				getProject().getDependencies().module("net.minecraft:minecraft:" + getJarVersionString("mapped")));
 	}
 
-	public void initFiles(MinecraftProvider minecraftProvider, MappingsProvider mappingsProvider) {
+	public void initFiles(MinecraftProviderImpl minecraftProvider, MappingsProviderImpl mappingsProvider) {
 		this.minecraftProvider = minecraftProvider;
 		minecraftIntermediaryJar = new File(getExtension().getUserCache(), "minecraft-" + getJarVersionString("intermediary") + ".jar");
 		minecraftMappedJar = new File(getJarDirectory(getExtension().getUserCache(), "mapped"), "minecraft-" + getJarVersionString("mapped") + ".jar");
@@ -159,7 +159,7 @@ public class MinecraftMappedProvider extends DependencyProvider {
 	}
 
 	protected String getJarVersionString(String type) {
-		return String.format("%s-%s-%s-%s", minecraftProvider.getMinecraftVersion(), type, getExtension().getMappingsProvider().mappingsName, getExtension().getMappingsProvider().mappingsVersion);
+		return String.format("%s-%s-%s-%s", minecraftProvider.minecraftVersion(), type, getExtension().getMappingsProvider().mappingsName, getExtension().getMappingsProvider().mappingsVersion);
 	}
 
 	public File getIntermediaryJar() {
