+++ src/main/java/net/fabricmc/loom/util/Constants.java
@@ -124,4 +124,13 @@ public class Constants {
 		private Knot() {
 		}
 	}
+
+	public static final class MappingNamespace {
+		public static final String OFFICIAL = "official";
+		public static final String NAMED = "named";
+		public static final String INTERMEDIARY = "intermediary";
+
+		private MappingNamespace() {
+		}
+	}
 }
